+++ .github/workflows/testcafe.yml
@@ -78,7 +78,7 @@ jobs:
         run: "curl localhost:8000/static/array.js"
 
       - name: Run chrome test
-        run: npx testcafe "browserstack:chrome" testcafe/e2e.ts --assertion-timeout 10000
+        run: npx testcafe "browserstack:chrome" testcafe/*.spec.js --assertion-timeout 10000
 
       - name: Run ie test
-        run: npx testcafe "browserstack:ie" testcafe/e2e.ts --assertion-timeout 10000
+        run: npx testcafe "browserstack:ie" testcafe/*.spec.js --assertion-timeout 10000
