+++ package.json
@@ -16,9 +16,10 @@
         "process-types": "mkdir -p dist; eslint src --ext .ts -c .eslintrc.ts.js --fix && tsc && cp -f src/*.d.ts dist/",
         "lint": "eslint src --fix",
         "prepublishOnly": "yarn lint && yarn test && yarn build && yarn test-react",
-        "test": "jest",
+        "test": "jest src",
         "test-react": "cd react; yarn test",
-        "test-watch": "jest --watch",
+        "test-watch": "jest --watch src",
+        "test-e2e": "testcafe ie testcafe/e2e.ts",
         "cypress": "cypress open"
     },
     "main": "dist/module.js",
